backend/requirements.txt
google-cloud-bigquery
pandas
db-dtypes
fastapi
openpyxl
uvicorn